+++ models/db/User.php
@@ -39,6 +39,7 @@ use yii\web\IdentityInterface;
  * @property null|MotionComment[] $motionComments
  * @property null|MotionSupporter[] $motionSupports
  * @property Site[] $adminSites
+ * @property Consultation[] $adminConsultations
  * @property ConsultationUserPrivilege[] $consultationPrivileges
  * @property ConsultationLog[] $logEntries
  * @property UserNotification[] $notifications
@@ -56,6 +57,7 @@ class User extends ActiveRecord implements IdentityInterface
     const PRIVILEGE_SCREENING                 = 3;
     const PRIVILEGE_MOTION_EDIT               = 4;
     const PRIVILEGE_CREATE_MOTIONS_FOR_OTHERS = 5;
+    const PRIVILAGE_ADMIN_USERS               = 6;
 
     /**
      * @return string[]
@@ -176,6 +178,15 @@ class User extends ActiveRecord implements IdentityInterface
     public function getAdminSites()
     {
         return $this->hasMany(Site::class, ['id' => 'siteId'])->viaTable('siteAdmin', ['userId' => 'id']);
+    }
+
+    /**
+     * @return \yii\db\ActiveQuery
+     */
+    public function getAdminConsultations()
+    {
+        return $this->hasMany(Consultation::class, ['id' => 'consultationId'])
+            ->viaTable('consultationAdmin', ['userId' => 'id']);
     }
 
     /**
@@ -585,12 +596,25 @@ class User extends ActiveRecord implements IdentityInterface
         if (in_array($this->id, $params->adminUserIds)) {
             return true;
         }
-        // @TODO Respect privilege table
+
         foreach ($consultation->site->admins as $admin) {
             if ($admin->id == $this->id) {
                 return true;
             }
         }
+
+        // Only site adminitrators are allowed to administer users.
+        // All other rights are granted to every consultation-level administrator
+        if ($privilege == User::PRIVILAGE_ADMIN_USERS) {
+            return false;
+        }
+
+        foreach ($consultation->admins as $admin) {
+            if ($admin->id == $this->id) {
+                return true;
+            }
+        }
+
         return false;
     }